+++ R/alluvial-data.r
@@ -45,7 +45,7 @@
 #'   indicating the type of alluvial structure ("none", "lodes", or "alluvia")
 #' @param silent Whether to print warning messages.
 #' @param key,value,id Numeric or character; the fields of \code{data}
-#'   corresponding to the axis (variable), stratum (value), and alluvium
+#'   corresponding to the axis (key), stratum (value), and alluvium
 #'   (identifying) variables.
 #' @param axes Numeric or character vector; the field(s) of \code{data}
 #'   corresponding to the axi(e)s (variable(s)).
@@ -62,6 +62,11 @@
 #'   \code{distill} accepts the character values \code{"first"} (used if
 #'   \code{distill} is \code{TRUE}), \code{"last"}, and \code{"most"} (which
 #'   returns the modal value).
+#' @param discern Logical value indicating whether to suffix values 
+#'   of the variables passed to \code{axes} that appear at more than one
+#'   variable in order to distinguish their factor levels. This forces the
+#'   levels of the combined factor variable \code{value} to be in the order of
+#'   the \code{axes}.
 #' @example inst/examples/ex-alluvial-data.r
 #' @export
 is_alluvial <- function(data, ..., logical = TRUE, silent = FALSE) {
@@ -145,7 +150,7 @@ is_alluvial_alluvia <- function(data,
 #' @export
 to_lodes <- function(data,
                      key = "x", value = "stratum", id = "alluvium",
-                     axes, diffuse = FALSE) {
+                     axes, diffuse = FALSE, discern = FALSE) {
   
   stopifnot(is_alluvial(data, axes = axes, silent = TRUE))
   
@@ -160,8 +165,20 @@ to_lodes <- function(data,
       stop("All 'diffuse' variables must be 'axes' variables.")
     }
   }
-  strata <- unique(unlist(lapply(lapply(data[axes], as.factor), levels)))
   
+  # combine factor levels
+  cat_levels <- unname(unlist(lapply(lapply(data[axes], as.factor), levels)))
+  if (any(duplicated(cat_levels)) & is.null(discern)) {
+    warning("Some strata appear at multiple axes.")
+  }
+  if (isTRUE(discern)) {
+    data <- discern_data(data, axes)
+    strata <- make.unique(unname(cat_levels))
+  } else {
+    strata <- unique(unname(cat_levels))
+  }
+  
+  # format data in preparation for 'gather()'
   data[[id]] <- 1:nrow(data)
   if (!is.null(diffuse)) diffuse_data <- data[, c(id, diffuse), drop = FALSE]
   for (i in axes) data[[i]] <- as.character(data[[i]])
@@ -171,6 +188,7 @@ to_lodes <- function(data,
                        rlang::UQ(axes),
                        factor_key = TRUE)
   res[[value]] <- factor(res[[value]], levels = strata)
+  
   if (!is.null(diffuse)) {
     res <- merge(res, diffuse_data, by = id, all.x = TRUE, all.y = FALSE)
   }
@@ -245,4 +263,18 @@ first <- dplyr::first
 last <- dplyr::last
 most <- function(x) {
   x[which(factor(x) == names(which.max(table(factor(x)))))[1]]
+}
+
+# 
+discern_data <- function(data, axes, sep = ".") {
+  list_levels <- lapply(lapply(data[axes], as.factor), levels)
+  cat_levels <- unlist(list_levels)
+  new_levels <- make.unique(unname(cat_levels))
+  i_levels <- cumsum(c(0, sapply(list_levels, length)))
+  for (i in seq_along(axes)) {
+    axis_levels <- as.numeric(as.factor(data[[axes[i]]]))
+    level_inds <- (i_levels[i] + 1):i_levels[i + 1]
+    data[[axes[i]]] <- new_levels[level_inds][axis_levels]
+  }
+  data
 }