+++ inst/examples/ex-alluvial-data.r
@@ -45,3 +45,22 @@ majors_alluvia2 <- to_alluvia(
   distill = "most"
 )
 head(majors_alluvia2)
+
+# options to distinguish strata at different axes
+gg <- ggplot(majors_alluvia,
+             aes(axis1 = CURR1, axis2 = CURR7, axis3 = CURR13))
+gg +
+  geom_alluvium(aes(fill = as.factor(student)), discern = TRUE) +
+  geom_stratum(discern = TRUE) +
+  geom_text(stat = "stratum", discern = TRUE, label.strata = TRUE)
+gg +
+  geom_alluvium(aes(fill = as.factor(student)), discern = FALSE) +
+  geom_stratum(discern = FALSE) +
+  geom_text(stat = "stratum", discern = FALSE, label.strata = TRUE)
+# warning when inappropriate
+ggplot(majors[majors$semester %in% paste0("CURR", c(1, 7, 13)), ],
+       aes(x = semester, stratum = curriculum, alluvium = student,
+           label = curriculum)) +
+  geom_alluvium(aes(fill = as.factor(student)), discern = TRUE) +
+  geom_stratum(discern = TRUE) +
+  geom_text(stat = "stratum", discern = TRUE)